Updating setup

+++ vars/setupStage.groovy
@@ -2,7 +2,7 @@
 
 def call(String stageName = 'Setup', Boolean cleanupWorkspace = true, Closure body) {
     stage(stageName) {
-        if (cleanupWorkspace) {
+        if (cleanupWorkspace && new File("${WORKSPACE}").exists()) {
             sh "rm -rf ${WORKSPACE}"
             sh "mkdir ${WORKSPACE}"
         }
